% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calib}
\alias{calib}
\title{Geolocator data recording sunrise and sunset}
\format{A data frame with 2,160 rows and 5 variables:
\describe{
  \item{lon}{Longitude}
  \item{lat}{Latitude}
  \item{date}{Date}
  \item{time}{Time of the measurement}
  \item{light}{Maximum light levels (lux) detected within the last two
  minutes}
}}
\source{
\url{https://doi.org/10.1111/ibi.12206}
}
\usage{
calib
}
\description{
A dataset containing raw geolocator light levels recording sunset and sunrise
events prior to placement on an animal
}
\keyword{datasets}
