% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavityuse-pkg.R
\docType{package}
\name{cavityuse-package}
\alias{cavityuse-package}
\alias{cavityuse}
\title{An R package for detecting cavity use from geolocator data}
\description{
\code{cavityuse} is an R package for detecting use of cavities (caves,
burrows, etc.) from geolocator data in a variety of animal species.
}
\details{
There are two main aspects of this package:

\enumerate{
  \item Detect patterns of cavity use in your data
  \itemize{
    \item \code{\link{sun_detect}()} Detect sunrise/sunset events in your
    data
    \item \code{\link{cavity_detect}()} Detect patterns of cavity use in your
    data (sunrise/sunset times from \code{\link{sun_detect}()} are used to
    detect cavity use at night).
    \item \code{\link{cavity_split}()} Split bouts of cavity use according
    the time frame you're interested in (i.e., midnight, noon, sunrise/sunset)
    }

\item Plot patterns of cavity use
 \itemize{
   \item \code{\link{cavity_plot}()}
   }
 }

We also include several practice data sets:

\itemize{
   \item \code{\link{flicker}}
   \item \code{\link{flicker_mult}}
   \item \code{\link{wtsp}}
   \item \code{\link{calib}}
   }
}
