% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flicker_mult}
\alias{flicker_mult}
\title{Geolocator data for two flicker individuals}
\format{A data frame with 13,680 rows and 6 variables:
\describe{
  \item{id}{Identity of the flicker}
  \item{lon}{Longitude}
  \item{lat}{Latitude}
  \item{date}{Date}
  \item{time}{Time of the measurement}
  \item{light}{Maximum light levels (lux) detected within the last two
  minutes}
}}
\source{
\url{https://doi.org/10.1111/ibi.12206}
}
\usage{
flicker_mult
}
\description{
A dataset containing raw geolocator light levels for several days for two
flicker individuals.
}
\keyword{datasets}
