% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.R
\name{sun_times}
\alias{sun_times}
\title{Get sunrise/sunset times}
\usage{
sun_times(loc, date, tz, angle = 6, interval = "none", type = "riseset")
}
\arguments{
\item{loc}{Vector/Data frame. Longitude and Latitude coordinates for location
of sun rise/set}

\item{date}{Vector. Date(s) to calculate sun rise/set for.}

\item{tz}{Timezone of the dates.}

\item{angle}{Numeric. For type = "riseset" the angle of the sun (6 = civil
twilight)}

\item{interval}{Character. How to organize the sunrise/sunset times}

\item{type}{Character. What kind of sunrise/sunset to calculate}
}
\description{
Calculate times of sunrise and sunset depending on the location and the date.
}
