% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tz_apply_offset}
\alias{tz_apply_offset}
\title{Apply tz offset without changing timezone}
\usage{
tz_apply_offset(data, tz_offset, cols = "time")
}
\arguments{
\item{data}{Data frame with "col" column}

\item{tz_offset}{Numeric. Number of hours for the offset}

\item{cols}{Character vector. Time columns over which to apply the offset

Applies a straight time shift to time columns and records this as an additional
column `tz_offset`. If a `tz_offset` has already been applied, it is first
removed.}
}
\description{
Apply tz offset without changing timezone
}
