% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tz_find_offset}
\alias{tz_find_offset}
\title{Use location to find TZ offset}
\usage{
tz_find_offset(lon, lat)
}
\arguments{
\item{lon}{Numeric. Decimal degree longitude}

\item{lat}{Numeric. Decimal degree latitude}
}
\description{
Uses location to establish what the non-daylight-savings offset would be
from UTC.
}
