% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tz_offset}
\alias{tz_offset}
\title{Calculate non-daylight savings timezone offset from UTC}
\usage{
tz_offset(tz)
}
\arguments{
\item{tz}{Character. Timezone to calculate offset from}
}
\description{
Calculate non-daylight savings timezone offset from UTC
}
