% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wtsp}
\alias{wtsp}
\title{Geolocator data for single white-throated sparrow individual}
\format{A data frame with 1,152 rows and 6 variables:
\describe{
  \item{id}{Identity of the flicker}
  \item{lon}{Longitude}
  \item{lat}{Latitude}
  \item{date}{Date}
  \item{time}{Time of the measurement}
  \item{light}{Maximum light levels (lux) detected within the last five
  minutes}
}}
\source{
\url{https://doi.org/10.1111/ibi.12206}
}
\usage{
wtsp
}
\description{
A dataset containing raw geolocator light levels for several days for a
single white-throated sparrow individual.
}
\keyword{datasets}
